pyproject.toml
[tool.poetry]
name = "example task template"
version = "0.1.0"
description = ""
authors = ["Daniil-Solo <daniil.solo1723@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.4"
pytest-env = "^1.1.5"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.25.1"
requests = "^2.32.3"
click = "^8.1.8"
pygithub = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
testpaths = ["src/tests"]
python_files = "test_*.py"
addopts = "-v"

[tool.pytest.ini_options]
asyncio_mode = "auto"
env_files = [".test.env"]

[tool.pytest_env]
SECOND_VALUE = "second"
